# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión Académica",
        "description": "API REST para la gestión de cursos, estudiantes, docentes y materias",
        "contact": {
            "name": "Soporte API",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/docentes": {
            "get": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Obtener lista de docentes activos",
                "operationId": "b588c11a6c17022729286fea84c0594f",
                "responses": {
                    "200": {
                        "description": "Lista de docentes obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Docentes obtenidos correctamente"
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_docente": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Juan"
                                                    },
                                                    "apellido": {
                                                        "type": "string",
                                                        "example": "Pérez"
                                                    },
                                                    "dni": {
                                                        "type": "string",
                                                        "example": "12345678"
                                                    },
                                                    "titulo_academico": {
                                                        "type": "string",
                                                        "example": "Licenciado en Matemáticas"
                                                    },
                                                    "estado": {
                                                        "type": "string",
                                                        "example": "A"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron docentes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Crear un nuevo docente",
                "operationId": "4bf41f182165401ab24053533defecc0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellido",
                                    "dni",
                                    "titulo_academico",
                                    "estado"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "titulo_academico": {
                                        "type": "string",
                                        "example": "Licenciado en Matemáticas"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "I"
                                        ],
                                        "example": "A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Docente creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Docente creado correctamente"
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "properties": {
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Juan"
                                                },
                                                "apellido": {
                                                    "type": "string",
                                                    "example": "Pérez"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}